Enhancements to the processing of data received over the powerline
The cm11a has a buffer of 10 bytes to data received ofver the power
line. Most receptions will consist of 2 byte, a house/unit address byte
and a fucntion byte. Some controller will send these in separate
messages. This enhancement trye to put the two massages back together
and act on the resul.